modifications during part 5 exercises

+++ Part4/bloglist/controllers/blogs.js
@@ -41,18 +41,31 @@ blogsRouter.post('/', async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: user._id
+    user: user.id
   })
 
   const newBlog = await blog.save()
+  await newBlog.populate('user', {username:1, name:1})
   user.blogs = user.blogs.concat(newBlog._id)
   await user.save()
   response.status(201).json(newBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+ 
+  if(user.id === blog.user.toString()){
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+  }
+  else{
+    return response.status(401).json({error: 'Unauthorized for this action'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -64,7 +77,10 @@ blogsRouter.put('/:id', async (request, response) => {
     url: body.url,
     likes: body.likes
   }
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {likes: body.likes})
+
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
+  .populate('user')
+  
   return response.json(updatedBlog)
 })
